--- a/.eslintrc
+++ b/.eslintrc
@@ -16,30 +16,40 @@
     "jsx-a11y/no-static-element-interactions": 0,
     "no-param-reassign": 0,
     "no-plusplus": 0,
-    "no-use-before-define": "off",
-    "react/prop-types": 0,
+    "no-use-before-define": 0,
+    "react/function-component-definition": [2, { "namedComponents": "arrow-function" }],
+    "react/jsx-filename-extension": [2, { "extensions": [".js", ".jsx", ".ts", ".tsx"] }],
+    "react/jsx-no-constructed-context-values": 0,
     "react/jsx-props-no-spreading": 0,
-    "react/jsx-filename-extension": [2, { "extensions": [".js", ".jsx", ".ts", ".tsx"] }]
+    "react/prop-types": 0
   },
   "plugins": ["@typescript-eslint", "import", "react"],
   "extends": [
     "airbnb",
     "eslint:recommended",
     "plugin:@typescript-eslint/eslint-recommended",
-    "plugin:@typescript-eslint/recommended",
     "plugin:@typescript-eslint/recommended-requiring-type-checking",
+    "plugin:@typescript-eslint/recommended",
     "plugin:react/recommended"
   ],
-  "ignorePatterns": ["src/stories", "**/*.test.js"],
+  "ignorePatterns": [
+    "**/*.test.*",
+    "dist",
+    "test.setup.js",
+    "vite.config.js"
+  ],
   "settings": {
     "import/resolver": {
-      "webpack": {
-        "moduleDirectory": ["src", "types"],
-        "extensions": [".ts"]
+      "alias": {
+        "map": [
+          ["@", "./src"],
+          ["types", "./types/types"]
+        ],
+        "extensions": [".ts", ".tsx"]
       }
     },
     "react": {
-      "version": "16.9.0"
+      "version": "16.14.0"
     }
   }
 }