--- a/.gitignore
+++ b/.gitignore
@@ -1,4 +1,5 @@
 # build
+/build
 /dist
 
 # dependencies