--- a/.storybook/config.js
+++ b/.storybook/config.js
@@ -1,3 +1,0 @@
-import { configure } from '@storybook/react'
-
-configure(require.context('../src', true, /\.stories\.js$/), module)