--- a/.storybook/main.js
+++ b/.storybook/main.js
@@ -1,3 +1,0 @@
-module.exports = {
-  addons: ['@storybook/addon-knobs'],
-};