--- a/src/Element/index.ts
+++ b/src/Element/index.ts
@@ -1,1 +1,3 @@
-export { default } from './Element';
+import Element from './Element';
+
+export default Element;