--- a/src/ElementsProvider/index.ts
+++ b/src/ElementsProvider/index.ts
@@ -1,1 +1,4 @@
-export { default, useElements } from './ElementsProvider';
+import ElementsProvider from './ElementsProvider';
+
+export default ElementsProvider;
+export { useElements } from './ElementsProvider';