--- a/src/Select/context/index.ts
+++ b/src/Select/context/index.ts
@@ -1,2 +1,4 @@
-export { default } from './SelectProvider';
+import SelectProvider from './SelectProvider';
+
+export default SelectProvider;
 export { default as useSelect } from './useSelect';