--- a/src/Select/hooks/useBoundary.ts
+++ b/src/Select/hooks/useBoundary.ts
@@ -3,9 +3,9 @@ import {
 } from 'react';
 
 import { Position } from 'types';
-import { usePanZoom } from 'context';
-import { onMouseUp, onMouseMove } from 'helpers/eventListener';
-import useContainerMouseDownPosition from 'hooks/useContainerMouseDownPosition';
+import { usePanZoom } from '@/context';
+import { onMouseUp, onMouseMove } from '@/helpers/eventListener';
+import useContainerMouseDownPosition from '@/hooks/useContainerMouseDownPosition';
 import { Boundary } from '../context/SelectContext';
 import { useSelect } from '../context';