--- a/src/Select/hooks/useBoundaryMove.ts
+++ b/src/Select/hooks/useBoundaryMove.ts
@@ -1,7 +1,7 @@
 import { useLayoutEffect } from 'react';
 
-import { onMouseDown, onMouseUp, onMouseMove } from 'helpers/eventListener';
-import { useElementMouseDownPosition, useElementMouseMovePosition } from 'hooks/useElementEventPosition';
+import { onMouseDown, onMouseUp, onMouseMove } from '@/helpers/eventListener';
+import { useElementMouseDownPosition, useElementMouseMovePosition } from '@/hooks/useElementEventPosition';
 import { MoveRef } from '../context/SelectContext';
 import { useSelect } from '../context';