--- a/src/Select/hooks/useBoundaryMove.ts
+++ b/src/Select/hooks/useBoundaryMove.ts
@@ -1,11 +1,14 @@
 import { useLayoutEffect } from 'react';
 
 import { onMouseDown, onMouseUp, onMouseMove } from '@/helpers/eventListener';
-import { useElementMouseDownPosition, useElementMouseMovePosition } from '@/hooks/useElementEventPosition';
+import {
+  useElementMouseDownPosition,
+  useElementMouseMovePosition,
+} from '@/hooks/useElementEventPosition';
 import { MoveRef } from '../context/SelectContext';
 import { useSelect } from '../context';
 
-type UseBoundaryMove = ({ grabElementsRef }: { grabElementsRef: MoveRef }) => void
+type UseBoundaryMove = ({ grabElementsRef }: { grabElementsRef: MoveRef }) => void;
 
 const useBoundaryMove: UseBoundaryMove = ({ grabElementsRef }) => {
   const {