--- a/src/Select/index.ts
+++ b/src/Select/index.ts
@@ -1,1 +1,3 @@
-export { default } from './Select';
+import Select from './Select';
+
+export default Select;