--- a/src/WithCover/WithCover.tsx
+++ b/src/WithCover/WithCover.tsx
@@ -7,8 +7,8 @@ import { ZOOM_INITIAL, ZOOM_MAX_DEFAULT } from '@/consts';
 import { CLASS_NAME } from '@/styles';
 import ElementsProvider from '@/ElementsProvider';
 import PanZoomProvider, { usePanZoom } from '@/context';
-import { PanZoom } from '../PanZoom';
 import produceStyle from '@/helpers/produceStyle';
+import { PanZoom } from '../PanZoom';
 
 type WithCoverProps = {
   cover: string,
@@ -77,30 +77,28 @@ const PanZoomWithCoverContext = (
   const [size, setSize] = useState<Size>({ width: '100%', height: '100%' });
 
   return (
-    <>
-      <PanZoomProvider
-        apiRef={apiRef}
-        {...props}
-        boundary
-        className={`${className || ''} ${CLASS_NAME}-with-cover`}
-        width={size.width}
-        height={size.height}
-        zoomInitial={scale}
-        zoomMin={scale}
-        zoomMax={zoomMax * scale}
-      >
-        <ElementsProvider>
-          <LoadCover
-            cover={cover}
-            onCoverLoad={onCoverLoad}
-            setScale={setScale}
-            setSize={setSize}
-          >
-            <PanZoom>{children}</PanZoom>
-          </LoadCover>
-        </ElementsProvider>
-      </PanZoomProvider>
-    </>
+    <PanZoomProvider
+      apiRef={apiRef}
+      {...props}
+      boundary
+      className={`${className || ''} ${CLASS_NAME}-with-cover`}
+      width={size.width}
+      height={size.height}
+      zoomInitial={scale}
+      zoomMin={scale}
+      zoomMax={zoomMax * scale}
+    >
+      <ElementsProvider>
+        <LoadCover
+          cover={cover}
+          onCoverLoad={onCoverLoad}
+          setScale={setScale}
+          setSize={setSize}
+        >
+          <PanZoom>{children}</PanZoom>
+        </LoadCover>
+      </ElementsProvider>
+    </PanZoomProvider>
   );
 };