--- a/src/WithCover/index.ts
+++ b/src/WithCover/index.ts
@@ -1,1 +1,3 @@
-export { default } from './WithCover';
+import WithCover from './WithCover';
+
+export default WithCover;