--- a/src/api/position.ts
+++ b/src/api/position.ts
@@ -4,25 +4,17 @@ import { Position, Zoom } from 'types';
 import positionClone from '@/helpers/positionClone';
 import produceStyle from '@/helpers/produceStyle';
 
-type GetPosition = (
-  props: { positionRef: MutableRefObject<Position> }
-) => () => Position
+type GetPosition = (props: { positionRef: MutableRefObject<Position> }) => () => Position;
 
 export const getPosition: GetPosition = ({ positionRef }) => () => positionClone(positionRef);
 
-type SetPosition = (
-  props: {
-    childRef: MutableRefObject<HTMLDivElement>,
-    positionRef: MutableRefObject<Position>,
-    zoomRef: Zoom,
-  }
-) => (x: number, y: number) => void
+type SetPosition = (props: {
+  childRef: MutableRefObject<HTMLDivElement>;
+  positionRef: MutableRefObject<Position>;
+  zoomRef: Zoom;
+}) => (x: number, y: number) => void;
 
-export const setPosition: SetPosition = ({
-  childRef,
-  positionRef,
-  zoomRef,
-}) => (x, y) => {
+export const setPosition: SetPosition = ({ childRef, positionRef, zoomRef }) => (x, y) => {
   const ref = childRef;
   const position = positionRef;
   position.current = { x, y };