Hey @sasza2, how are you? It's been a while since the last PR. I hope you will like this one.

# Overview

In this PR, I've made a global refactor of the whole stack to let us use the latest technologies on the web during development:

- Webpack > [Vite](https://vitejs.dev/)
- Jest > [Vitest](https://vitest.dev/)
- Storybook > [Ladle](https://ladle.dev/)
- npm > [pnpm](https://pnpm.io/)
- React 16 > React 18

## Why Vite?

Because Vite is extremely fast and very much simpler to setup, compared to Webpack.
Also, with this setup, **the bundle size of React Panzoom is halved**: from 48K to 24K.

## Why Vitest?

Vitest supports TypeScript by default without any workaround to set it up (unlike Jest).
No need to learn its API since it is like the Jest one (describe, it, etc.).
In performance, Vitest is slightly better than Jest (and in some cases TS-Jest can be extremely slow).

## Why Ladle?

Storybook is extremely heavy in its dependencies and very slow to run.
With this setup, **the size of the node_modules folder dropped** from 515M to 241M.
You can see a comparison with its newest counterpart [here](https://blog.logrocket.com/ladle-storybook-performance-project-sizes).
Its API is similar to Storybook, but I've found it simpler to set up (no configuration).

## Why pnpm?

pnpm is much faster than npm, with and without caching scenarios.
In my case, the time of the installation dropped from:
- 10s to 3.6s (with lock file)
- 19s to 37s (without)

## Why React 18?

Well, it's just updating. React 16 was released 5 years ago, and the latest versions of the other dependencies easily require versions higher than React 16 these days.

# More on this PR

- few bug fixes
- enforced eslint rules while refactoring
- tests pass ✅
- rewritten stories & tests in TypeScript
- absolute paths via the "@" convention
- we now have 13 dev dependencies instead of 29
- global formatting una-tantum (via Prettier, using eslint coding standards)