@HighLiuk 2022-12-21
@sasza2 please let me know what do you think of them all. It may seem very disruptive, but it's probably not. Have you ever tried some of the stack techs I mentioned?

P.S. No idea if the npm publishing will be Ok, I guess that's up to you to test. However I've tried to launch the CI commands with Node 14 but it failed miserably 😂 by googling it seemed to be related to Node 14 so I've set Node 16 to launch the tests in CI.