@sasza2 2022-12-21
Hi! you're really awesome @HighLiuk  :D it's so nice for me to see this PR.

Vite, Vitest, Ladle, pnmp - ok

about updating React peer dependency - I think it should stay with current version. Please, look at https://www.npmjs.com/package/react?activeTab=versions
most downloads are under version 18 and package itself is not using any feature above version 16.8 (hooks). Every project using this library could install any React version above 16.8 and it should work. Unless react-panzoom is not using features from newest version like useId / useTransition / useDeferredValue etc I don't see the point for changing this.

```
Well, it's just updating. React 16 was released 5 years ago, and the latest versions of the other dependencies easily require versions higher than React 16 these days.
```

Ok, if this is the case, we should figure out how to leave it in old version. Could you name what dependencies require newest version? Thanks.

```
Have you ever tried some of the stack techs I mentioned?
```

Yes, I did and I think you did great job here for upgrading current stack here with latest tech

develop now is default branch, because I've added workflow on pushing to master. On every push workflow is building and publishing library  (I wanted to have control over releasing new versions). If you could point everything to branch sasza2:develop it would be great, because maybe I will add some changes before merging it to -> develop and later -> master

`I've set Node 16 to launch the tests in CI.`

okay, I'll check that

`few bug fixes`

Could you say something more about this? thanks