@HighLiuk 2022-12-21
Thanks @sasza2 for all the support. Let's break it down so we can merge this! I'll change the target branch as soon as I can.

> few bug fixes

Check these two commits:
- Fix Naming Collision
- ElementProps x and y should be optional

For React 16, as far as I remember there were two dependencies that give problems: Testing Library & Ladle, but mostly Testing Library since I had dependencies conflicts even on Testing Library 10, which should be the current version at origin master... Not so sure about that, but you can investigate it further by a few `pnpm why react-dom`