@HighLiuk 2022-12-21
Hey @sasza2, I've looked at https://www.npmjs.com/package/react?activeTab=versions and I've done even more.
Look at this:

# React Weekly Downloads segmented by version

```mermaid
pie showData title By Major Release
"17":6077758
"16":5672055
"18":4997284
"15":360262
"Other":160137
```

```mermaid
pie showData title By Minor Release
"17.0":6077758
"18.2":4113997
"16.14":3579091
"16.13":1124089
"18.1":627715
"16.8":295467
"18.0":250010
"16.12":239582
"15.7":207205
"Other":752582
```

```mermaid
pie showData title By Patch Release
"17.0.2":5570062
"18.2.0":4113997
"16.14.0":3579091
"16.13.1":1080795
"18.1.0":627715
"17.0.1":479759
"18.0.0":250010
"16.12.0":239582
"15.7.0":207205
"16.8.6":194754
"Other":924526
```

It looks like 16.13 is the ideal minimum version we can adopt if we are able to. What do you think about?

P.S. source code to generate the data above comes from [this gist](https://gist.github.com/HighLiuk/fe203e7bffc78c0615a9bb693f9984d5) 😉