@sasza2 2022-12-22
okay, thanks for graphics. I still don't get it, why do you want to upgrade React version so much hehe?:D I think if library is not using any new features from new versions of React it should stay lower like for example:

https://github.com/react-grid-layout/react-grid-layout/blob/master/package.json#L90
https://github.com/react-bootstrap/react-bootstrap/blob/master/package.json#L150
https://github.com/styled-components/styled-components/blob/main/packages/styled-components/package.json#L79
https://github.com/react-dnd/react-dnd/blob/main/packages/react-dnd/package.json#L50

do we have any argument for updating this version except that now its not working with Testing Library & Ladle? I think we can add newest version of React to devDependencies and it should work