@HighLiuk 2022-12-22
Haha yeah I like updating stuff (otherwise I wouldn't have embarked in such a refactor lol)
Anyways, 99% that's enough, though having the same version of React on both devDependencies and peerDependencies lets us ensure that if I launch the tests on my local environment and the tests pass, we increase the confidence we have in our code when a developer install the library with the same version in peerDeps.
Not so sure if I've explained myself well