@sasza2 2022-12-22
you're right. ok, so 16.13 it the lowest version for Testing Library & Ladle to run? If so, we can stay with this version for now and if this PR goes to develop I will try run 16.9 hehe, maybe I figure something out