@sasza2 2022-12-22
ok, I was able to hack tests to "work", but honestly I hate this idea. so for now I think this - peerDependency: 16.9 and devDependency: lowest version to pass tests without hacks. I know that versions will be different, but I think it can stay for now