@HighLiuk 2022-12-22
Hey @sasza2 I've changed the target branch to develop.
Also, I've managed to install all dependencies without conflicts with:

- @ladle/react latest version
- @testing-library/react v12.1.4
- react & react-dom v16.14.0

Unfortunately Ladle requires React 16.4 since it was born, while the same version of Testing Library works with any React version (and it is the last version of Testing Library with that property, otherwise it requires React 18).

Test pass, though there is a bug with the WithCover component, but I guess it has nothing to do with upgrading the stack. You can reproduce it by `pnpm dev` and going to `http://localhost:61000/?story=cover--cover`

What do you think of the PR? There are three possibilities:

1. It's ok, merge it
2. We still want React 16.9, so we can try working locally with Ladle even with React 16.9, and hope that no bugs are encountered - but in that case, it's just for dev. That way we install 16.9 and test against that version
3. We switch back Ladle to Storybook - sigh - so we can work with no conflicts and version 16.9 of React

What's your choice?