@HighLiuk 2022-12-23
Here we go @sasza2, I've completed the rebase and I've also synced it with the most recent version of branch develop.

- ✅ Upgrade Stack to Vite, Vitest, Ladle, pnpm
- ✅ Tests Pass
- ✅ Lint Pass
- ✅ Build Pass
- ✅ No Dependencies Conflicts
- ✅ React >=16.14
- ✅ Git History clean
- ✅ Fix minor bugs (see [here](https://github.com/sasza2/react-panzoom/pull/72/commits/15590778252f17334898cf1fd7ad0243d3fd72d2) and [there](https://github.com/sasza2/react-panzoom/pull/72/commits/fb2ea885feb2b6b357be97094ba608df914fcae6))

Thanks in advance 😉👍🏼